% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimxOptimPackage.r
\name{SARALpred}
\alias{SARALpred}
\title{Genetic Risk Prediction Using a Spatial Autoregressive Model with Adaptive Lasso}
\usage{
SARALpred(Y, Geno, TrainID, genelist = list(), Similarity = NULL,
  SaveSimilarity = FALSE, YTest = NULL, compare = "rrblup",
  assess = TRUE, methods = c("Cp", "AICc", "GCV", "BIC", "MML"))
}
\arguments{
\item{Y}{A vector of phenotypes. Missing values should be coded as \code{NA}.}

\item{Geno}{A matrix of genotypes, with each row representing individual and column representing single nucleotide variant.}

\item{TrainID}{A bool vector with \code{true} indicating the individual is in the training sample.}

\item{genelist}{A list with each element represents a group.}

\item{Similarity}{A list with \code{i}-th element measuring the similarity for a particular chuck of genome listed in the \code{genelist[[i]]}.}

\item{SaveSimilarity}{A bool indicating whether the similarity should be saved for futher use. By default, similarity is not saved.}

\item{compare}{A string that indicates which method the SARAL compares to. Currently it only supports GBLUP.}

\item{assess}{A bool. If true, the method will calculate Pearson correlation and mean square error from the testing data. This option is ignored if all the phenotypes of the testing sample is \code{NA} (i.e. \code{sum(!is.na(Y[!TrainID]))==0}).}

\item{methods}{A string vector that determines which model selection criterion is used to select model for SARAL. By default, it uses Mallows' Cp (Cp), bias corrected AIC (AICc), generalized cross validataion (GCV), BIC, and information theoretic minimum message length (MML).}
}
\value{
A list that contains the predicted values, the pearson correlations and mean square errors (if \code{assess=TRUE}), and error if algorithm does not converge. The list also contains a list of similarity measure (if \code{SaveSimilarity=TRUE}).
}
\description{
\code{SARpred} returns the predicted values of the phenotypes.
}
\details{
This function is used to predict outcomes based on individual's genetic profiles.
}
\examples{
# This may take 2 hours, don't run if not needed #
Y=SARAL::example_SARAL$Y;
Geno=SARAL::example_SARAL$Geno;
TrainID=SARAL::example_SARAL$TrainID;
genelist=SARAL::example_SARAL$Genelist;
Result<-SARALpred(Y,Geno,TrainID,genelist,compare='rrblup',assess=TRUE,methods=c("Cp","AICc","GCV","BIC","MML"))
}
